Add support for controlling the size of the string associated to each match.
YARA maintains in memory a copy of each matched string, this can consume a lot of memory if we copy the full matching strings in all cases. For that reason YARA only stores the first 512 bytes of each matching string by default, but this amount can be controlled with a call to yr_set_configuration(YR_CONFIG_MAX_MATCH_DATA, value). With this commit we expose this API to Go users.